+++ src/main/java/subway/domain/line/model/LineRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
